Staff Apps to separate application
Moved all models/views/templates, etc related to Staff Applications to own app named staffapps

Still some refining to do. Currenty, I am not passing form POST data if a form is invalid back to the page, so users would have to re-type entire app if something was wrong. Since the page simply redirects on invalid post and uses django messages to transfer invalid form messages, I will likely use sessions to store and transfer the form POST data. Not 100% sure how that will go currently.
