+++ raptorWeb/staffapps/admin.py
@@ -1,23 +1,20 @@
 from django.contrib import admin
 from django.forms import TextInput, Textarea
 from django.db import models
-from raptormc.models import Server, InformativeText, PlayerCount, PlayerName, UserProfileInfo, DiscordUserInfo
 
-class ServerAdmin(admin.ModelAdmin):
+from staffapps.models import ModeratorApplication, AdminApplication
+
+class AdminApplicationAdmin(admin.ModelAdmin):
     formfield_overrides = {
         models.CharField: {'widget': TextInput(attrs={'size':'35'})},
         models.TextField: {'widget': Textarea(attrs={'rows':6, 'cols':60})},
     }
 
-class InformativeTextAdmin(admin.ModelAdmin):
+class ModApplicationAdmin(admin.ModelAdmin):
     formfield_overrides = {
         models.CharField: {'widget': TextInput(attrs={'size':'35'})},
         models.TextField: {'widget': Textarea(attrs={'rows':6, 'cols':60})},
     }
 
-admin.site.register(Server, ServerAdmin)
-admin.site.register(PlayerCount)
-admin.site.register(PlayerName)
-admin.site.register(InformativeText, InformativeTextAdmin)
-admin.site.register(UserProfileInfo)
-admin.site.register(DiscordUserInfo)
+admin.site.register(AdminApplication, AdminApplicationAdmin)
+admin.site.register(ModeratorApplication, ModApplicationAdmin)
