--- raptorWeb/templates/raptormc/applications/modapp.html
@@ -1,42 +1,0 @@
-{% extends 'raptormc/base.html' %}
-{% load django_bootstrap5 %}
-
-{% block page_title %}
-<title>
-  
-  ShadowRaptor | Moderator Application
-  
-</title>
-{% endblock %}
-{% block metadata %}
-<meta property="og:title" content="ShadowRaptor | Moderator Application" />
-<meta property="og:description" content="This page contains our Moderator Application." />
-<meta property="og:url" content="https://shadowraptor.net/applications/mod" />
-{% endblock %}
-
-{% block main_block %}
-<section id="welcome" class="container">
-
-  <header class="card bg-dark opacity-75 text-white p-4 mt-4">
-    <h3 class="pageHeader fw-bold text-center">Moderator Application</h3>
-  </header>
-
-</section>
-
-<section>
-
-  <div class="container card opacity-75 mt-4 p-3 border">
-
-    <form class="form p-3" method="POST">
-
-      {% csrf_token %}  
-      {% bootstrap_form modform %}
-      {% bootstrap_button button_type="submit" content="Submit Application" %}
-      {% bootstrap_button button_type="reset" content="Reset" %}
-
-    </form>
-
-  </div>
-
-</section>
-{% endblock %}
